reorganize and lint
