+++ spotlight-client/src/contentModels/__fixtures__/tenant_content_exhaustive.ts
@@ -141,6 +141,16 @@ const content: DeepRequired<TenantContent> = {
       description: "test ParoleProgrammingCurrent description",
       methodology: "test ParoleProgrammingCurrent methodology",
     },
+    ParoleSuccessAggregate: {
+      name: "test ParoleSuccessAggregate name",
+      description: "test ParoleSuccessAggregate description",
+      methodology: "test ParoleSuccessAggregate methodology",
+    },
+    ProbationSuccessAggregate: {
+      name: "test ProbationSuccessAggregate name",
+      description: "test ProbationSuccessAggregate description",
+      methodology: "test ProbationSuccessAggregate methodology",
+    },
   },
 };
